Fix opencl runtime code format.

See merge request !317

+++ mace/core/runtime/opencl/opencl_runtime.h
@@ -46,9 +46,9 @@ class OpenCLRuntime {
   cl::CommandQueue &command_queue();
 
   void GetCallStats(const cl::Event &event, CallStats *stats);
-  uint32_t GetDeviceMaxWorkGroupSize();
-  uint32_t GetKernelMaxWorkGroupSize(const cl::Kernel &kernel);
-  uint32_t GetKernelWaveSize(const cl::Kernel &kernel);
+  uint64_t GetDeviceMaxWorkGroupSize();
+  uint64_t GetKernelMaxWorkGroupSize(const cl::Kernel &kernel);
+  uint64_t GetKernelWaveSize(const cl::Kernel &kernel);
   cl::Kernel BuildKernel(const std::string &program_name,
                          const std::string &kernel_name,
                          const std::set<std::string> &build_options);
