+++ mace/kernels/opencl/helper.cc
@@ -201,7 +201,8 @@ void TuningOrRun3DKernel(const cl::Kernel &kernel,
                          StatsFuture *future) {
   auto runtime = OpenCLRuntime::Global();
   auto params_generator = [&]() -> std::vector<std::vector<uint32_t>> {
-    const uint32_t kwg_size = runtime->GetKernelMaxWorkGroupSize(kernel);
+    const uint32_t kwg_size =
+        static_cast<uint32_t>(runtime->GetKernelMaxWorkGroupSize(kernel));
     std::vector<uint32_t> local_ws(3, 0);
     local_ws[0] = std::min<uint32_t>(gws[0], kwg_size);
     local_ws[1] = std::min<uint32_t>(gws[1], kwg_size / local_ws[0]);
@@ -304,7 +305,8 @@ void TuningOrRun2DKernel(const cl::Kernel &kernel,
                          StatsFuture *future) {
   auto runtime = OpenCLRuntime::Global();
   auto params_generator = [&]() -> std::vector<std::vector<uint32_t>> {
-    const uint32_t kwg_size = runtime->GetKernelMaxWorkGroupSize(kernel);
+    const uint32_t kwg_size =
+        static_cast<uint32_t>(runtime->GetKernelMaxWorkGroupSize(kernel));
     uint32_t local_ws[2];
     local_ws[0] = std::min<uint32_t>(gws[0], kwg_size);
     local_ws[1] = std::min<uint32_t>(gws[1], kwg_size / local_ws[0]);
